requirements.txt
tqdm~=4.66.1
colorlog~=6.8.2
mysql-connector-python~=8.3.0
Unidecode~=1.3.7